Write tests for remaining iteration2 methods
